fixed imports after tests moved
